+++ samples/net/coap_client/src/main.c
@@ -163,7 +163,7 @@ static int periodic_coap_request_loop(void)
 }
 
 static void l4_event_handler(struct net_mgmt_event_callback *cb,
-			     uint32_t event,
+			     uint64_t event,
 			     struct net_if *iface)
 {
 	switch (event) {
@@ -186,7 +186,7 @@ static void l4_event_handler(struct net_mgmt_event_callback *cb,
 	}
 }
 static void connectivity_event_handler(struct net_mgmt_event_callback *cb,
-						uint32_t event,
+						uint64_t event,
 						struct net_if *iface)
 {
 	if (event == NET_EVENT_CONN_IF_FATAL_ERROR) {
