+++ samples/wifi/thread_coex/src/ot_coex_functions.h
@@ -218,7 +218,7 @@ static struct net_mgmt_event_callback net_addr_mgmt_cb;
  *
  * @return No return value.
  */
-void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 	struct net_if *iface);
 
 /**
@@ -226,7 +226,7 @@ void wifi_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_e
  *
  * @return No return value.
  */
-void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint32_t mgmt_event,
+void net_mgmt_event_handler(struct net_mgmt_event_callback *cb, uint64_t mgmt_event,
 	struct net_if *iface);
 
 /**
